Refactor client ID types from int/Integer to Long
Updated all client-related controller, service, repository, and DTO methods to use Long instead of int/Integer for client IDs and related fields. Also removed unused fields from ClientResponseDTO and ClientMapper.

+++ src/main/java/com/solarize/solarizeWebBackend/modules/client/ClientRepository.java
@@ -11,8 +11,8 @@ public interface ClientRepository extends JpaRepository<Client, Long> {
     boolean existsByEmail(String email);
     boolean existsByPhone(String phone);
     boolean existsByCnpj(String cnpj);
-    boolean existsByDocumentNumberAndIdNot(String documentNumber, Integer id);
-    boolean existsByEmailAndIdNot(String email, Integer id);
-    boolean existsByPhoneAndIdNot(String phone, Integer id);
-    boolean existsByCnpjAndIdNot(String cnpj, Integer id);
+    boolean existsByDocumentNumberAndIdNot(String documentNumber, Long id);
+    boolean existsByEmailAndIdNot(String email, Long id);
+    boolean existsByPhoneAndIdNot(String phone, Long id);
+    boolean existsByCnpjAndIdNot(String cnpj, Long id);
 }
